Move db tests to file
